+++ google/CHANGELOG.md
@@ -4,7 +4,7 @@ This file documents changes to the `workbench-google` library, including notes o
 
 ## 0.21
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.21-296ce50"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.21-TRAVIS-REPLACE-ME"`
 
 ### Added
 
@@ -25,6 +25,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google" % "0.
 - Made `RetryPredicates` handle `null`s more safely
 - Creating a group sometimes returns a 5xx error code and leaves behind a partially created group which caused problems 
 when we retried creation. Changed to delete the partially created group before retrying
+- Cross build to scala 2.13
 
 ## 0.20